fix refactoring bugs
