+++ test/demo/utils/common.ts
@@ -1,6 +1,6 @@
 import {Identity} from 'openfin/_v2/main';
 
-import {Intent} from '../../../src/client/intents';
+import {Intent} from '../../../src/provider/intents';
 import {withTimeout} from '../../../src/provider/utils/async';
 import {testManagerIdentity, appStartupTime} from '../constants';
